% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_fiora.R
\name{install_fiora}
\alias{install_fiora}
\title{Install the python module `fiora` into a conda environment.}
\usage{
install_fiora(conda_name = "fiora")
}
\arguments{
\item{conda_name}{The name of the conda environment where fiora shall be installed to.}
}
\value{
A list providing the current OS and path information on the current
    python executable and the fiora script.
}
\description{
This function will check and perform the installation of three
    components in the following order: reticulate, miniconda and fiora. It
    will ensure that a working conda environment `fiora` is available. This
    is a prerequisite for both, \link{run_app} and \link{run_script}.
}
\examples{
\dontrun{
  # this will install packages and software on your machine
  install_fiora()
}

}
