% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_spec.R
\name{plot_spec}
\alias{plot_spec}
\title{Plot MS^2 spectrum.}
\usage{
plot_spec(s, show_neutral_losses = TRUE, show_smiles = TRUE, ...)
}
\arguments{
\item{s}{A valid spectrum as predicted by fiora.}

\item{show_neutral_losses}{Connect main peaks by gray lines and annotate the
respective sum formula of the neutral loss.}

\item{show_smiles}{Overlay SMILES (if present in columns of s) in plot.}

\item{...}{Passed on to `InterpretMSSpectrum::PlotSpec()`.}
}
\value{
Creates an annotated plot of a mass spectrum and returns the spectrum invisibly.
}
\description{
This is a wrapper around function `InterpretMSSpectrum::PlotSpec()`
    allowing to add chemical structures of MS^2 fragments to the plot.
}
\details{
See examples and documentation of `InterpretMSSpectrum::PlotSpec()`.
}
\examples{
fl <- system.file("extdata/annotated_output.mgf", package = "fioRa")
tmp <- fioRa::read_fiora(fl = fl)
s <- tmp[[3]][["spec"]]
plot_spec(s = s)
plot_spec(s = s, show_neutral_losses = FALSE)
plot_spec(s = s, masslab = 0.05)
}
\seealso{
[InterpretMSSpectrum::PlotSpec()]
}
