% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_predict.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(
  Name = "Example_0",
  SMILES = "CC1=CC(=O)OC2=CC(OS(O)(=O)=O)=CC=C12",
  Precursor_type = "[M-H]-",
  CE = 17,
  Instrument_type = "HCD"
)
}
\arguments{
\item{Name}{The compound name.}

\item{SMILES}{The SMILES code of the compound.}

\item{Precursor_type}{Precursor type.}

\item{CE}{Collision energy.}

\item{Instrument_type}{Instrument type.}
}
\value{
A list with the fiora results for the specified compound.
}
\description{
A wrapper around the python fiora model.
}
\examples{
predict()

}
