% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{The example set of test compounds provided with FIORA.}
\format{
A character vector of length = 12 containing the readLines equivalent
    of a fiora input file in csv format, with header, 10 example compounds,
    defined via their SMILES, and an empty final line (or vector element
    respectively).
}
\source{
https://github.com/BAMeScience/fiora/blob/main/examples/example_input.csv
}
\usage{
data(test_data)
}
\description{
The example set of test compounds provided with FIORA.
}
\keyword{datasets}
