% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{An example set of data collected for a CRM.}
\format{
A character vector of length = 12 containing the readLines equivalent
    of a fiora input file in csv format, with header, 10 example compounds,
    defined via their SMILES, and an empty final line (or vector element
    respectively).
}
\source{
jan.lisec@bam.de
}
\usage{
data(test_data)
}
\description{
An example set of data collected for a CRM.
}
\keyword{datasets}
